+++ clodoo/clodoo/tests/__pytest_transodoo.py
@@ -12,7 +12,7 @@
 # import pytest
 from clodoo import transodoo
 
-__version__ = '1.0.1'
+__version__ = '1.0.1.1'
 
 
 class TestClass: